remove some code
